requirements.txt
flask==3.1.1
requests==2.32.3
gunicorn==23.0.0
google-api-python-client==2.166.0
google-auth==2.38.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
